requirements.txt
customtkinter>=6.4.2
python-mpv>=1.6.0
Pillow>=9.0.0